+++ packages/core/src/lib/actions/selecting.action.ts
@@ -1,6 +1,5 @@
 import { BaseAction } from './base.action';
 import { DiagramModel } from '../models/diagram.model';
-import { Coords } from '../interfaces/coords.interface';
 
 export class SelectingAction extends BaseAction {
   mouseX2: number;
@@ -41,15 +40,27 @@ export class SelectingAction extends BaseAction {
     return this.dimensions;
   }
 
-  containsElement({ x, y }: Coords, diagramModel: DiagramModel): boolean {
+
+  containsElement(topLeftPoint: { x: number, y: number }, bottomRightPoint: { x: number, y: number }, diagramModel: DiagramModel): boolean {
     const z = diagramModel.getZoomLevel() / 100.0;
     const dimensions = this.getBoxDimensions();
 
-    return (
-      x * z + diagramModel.getOffsetX() > dimensions.left &&
-      x * z + diagramModel.getOffsetX() < dimensions.right &&
-      y * z + diagramModel.getOffsetY() > dimensions.top &&
-      y * z + diagramModel.getOffsetY() < dimensions.bottom
+    // check if box contain top left point
+    const isContainsTopLeftPoint = (
+      topLeftPoint.x * z + diagramModel.getOffsetX() > dimensions.left &&
+      topLeftPoint.x * z + diagramModel.getOffsetX() < dimensions.right &&
+      topLeftPoint.y * z + diagramModel.getOffsetY() > dimensions.top &&
+      topLeftPoint.y * z + diagramModel.getOffsetY() < dimensions.bottom);
+
+    // check if box contain bottom right point
+    const isContainBottomRightPoint = (
+      bottomRightPoint.x * z + diagramModel.getOffsetX() > dimensions.left &&
+      bottomRightPoint.x * z + diagramModel.getOffsetX() < dimensions.right &&
+      bottomRightPoint.y * z + diagramModel.getOffsetY() > dimensions.top &&
+      bottomRightPoint.y * z + diagramModel.getOffsetY() < dimensions.bottom
     );
+
+    // only if box contains both top left and bottom right points, the element is contains in the box
+    return isContainsTopLeftPoint && isContainBottomRightPoint;
   }
 }