+++ packages/core/src/lib/engine.core.ts
@@ -261,6 +261,29 @@ export class DiagramEngineCore {
     return model.getLocked();
   }
 
+   /**
+   * @description get node rectangle points (top left, top right, bottom left, bottom right)
+   */
+  getNodeRectPoints(node: NodeModel) {
+    if (!node) {
+      console.warn('[RxZu] No input node were found');
+      return;
+    }
+    
+    const width = node.getWidth();
+    const height = node.getHeight();
+
+    // get top left point of node
+    const { x, y } = node.getCoords();
+
+    return {
+      topLeft: { x, y },
+      topRight: { x: x + width, y },
+      bottomLeft: { x: x, y: y + height },
+      bottomRight: { x: x + width, y: y + height },
+    }
+  }
+
   /**
    * @description get the bounding rectangle of the input group of nodes
    * @param nodes the group of nodes to calculate the retcangle
@@ -286,16 +309,29 @@ export class DiagramEngineCore {
 
     // go over all nodes and calc the min, max points of all of them
     nodes.forEach(node => {
+
+      // get coords of current node
       const { x, y } = node.getCoords();
+
+      // get client width and client height of current node
+      const nodeElement = this.getNodeElement(node);
+      const { clientWidth, clientHeight } = nodeElement;
+
+      // if this node located at the leftmost point
       if (x < minX) minX = x;
-      if (x > maxX) {
-        maxX = x;
-        rightNodeWidth = node.getWidth();
-      }
+
+      // if this node located at the topmost point
       if (y < minY) minY = y;
-      if (y > maxY) {
+
+      // if this node located at the rightmost point including the width
+      if (x + clientWidth > maxX + rightNodeWidth) {
+        maxX = x;
+        rightNodeWidth = clientWidth;
+      }
+      // if this node located at the bottommost point including the height
+      if (y + clientHeight > maxY + bottomNodeHeight) {
         maxY = y;
-        bottomNodeHeight = node.getHeight();
+        bottomNodeHeight = clientHeight;
       }
     })